+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Chg #161: Allow to reuse manager test code in related packages (@arogachev)
 - New #161: Add `ManagerInterface` (@arogachev)
 - Chg #161: Raise PHP version to 8.0 (@arogachev)
+- Enh #165: Improve perfomance (@arogachev)
+- Bug #178: Exclude parent role from `Manager::getAllChildRoles()` (@arogachev)
 
 ## 1.0.2 April 20, 2023